DjangoRestframework/requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
Django==4.0.4
django-cors-headers==3.12.0
djangorestframework==3.13.1
pytz==2022.1
six==1.16.0
sqlparse==0.4.2

opencv-python~=3.4.10.35
numpy~=1.24.4
scipy~=1.10.1